carousal Added
Add Carousal from flowbite css and customize it and make responsive
