+++ src/App.tsx
@@ -1,19 +1,17 @@
 import { BrowserRouter, Route, Routes } from "react-router-dom";
 import "./App.css";
 import Homepage from "./components/Homepage";
+import Layout from "./components/Layout";
 
 function App() {
-
   return (
     <>
       <BrowserRouter>
-      <Routes>
-        <Route index element={
-          <Homepage/>
-        }>
-          
-        </Route>
-      </Routes>
+        <Routes>
+          <Route path="/" element={<Layout />}>
+            <Route index element={<Homepage />} />
+          </Route>
+        </Routes>
       </BrowserRouter>
     </>
   );
