+++ src/components/Layout.tsx
@@ -1,0 +1,14 @@
+
+import { Outlet } from "react-router-dom"
+import NavBar from "./NavBar/NavBar"
+
+const Layout = () => {
+  return (
+    <div>
+        <NavBar/>
+        <Outlet/>
+    </div>
+  )
+}
+
+export default Layout
