+++ src/components/NavBar/NavBar.css
@@ -1,0 +1,16 @@
+::-webkit-scrollbar {
+  width: 5px;
+}
+
+::-webkit-scrollbar-thumb {
+  background: #756aee;
+  border-radius: 5px;
+}
+
+.NavCarousal  [data-testid="carousel-indicator"] {
+  background-color: #e2e8f0;
+  border-radius: 0px ;
+}
+.NavCarousal [class="h-3 w-3 rounded-full bg-white dark:bg-gray-800"] {
+  background-color: #756aee;
+}
