+++ src/components/Slider/Slider.css
@@ -1,0 +1,49 @@
+.SliderContainer .indiana-scroll-container--hide-scrollbars {
+  border-radius: 0px !important;
+}
+.SliderContainer
+  [class="absolute bottom-5 left-1/2 flex -translate-x-1/2 space-x-3"] {
+  height: fit-content;
+  width: fit-content;
+  margin: 0px;
+  flex-direction: column;
+  gap: 20px;
+  top: 50%;
+  left: auto !important;
+  right: 10px !important;
+  bottom: 0px !important;
+}
+.SliderContainer [data-testid="carousel-indicator"] {
+  margin-left: 2px !important;
+  width: 40px !important;
+  height: 4px !important;
+  background-color: #ffffff !important;
+  border-radius: 0px !important;
+  position: relative;
+}
+.SliderContainer [data-testid="carousel-indicator"]::before {
+  content: "";
+  position: absolute;
+  left: 0;
+  top: 0;
+  height: 100%;
+  width: 100%;
+  transform-origin: left;
+  --tw-scale-x: 0;
+  transform: translate(var(--tw-translate-x), var(--tw-translate-y))
+    rotate(var(--tw-rotate)) skewX(var(--tw-skew-x)) skewY(var(--tw-skew-y))
+    scaleX(var(--tw-scale-x)) scaleY(var(--tw-scale-y));
+  background: #756aee;
+  transition-property: transform;
+  transition-timing-function: cubic-bezier(0.4, 0, 0.2, 1);
+  transition-duration: 300ms;
+}
+.SliderContainer [data-testid="carousel-indicator"]:hover::before {
+  --tw-scale-x: 1;
+  transform: translate(var(--tw-translate-x), var(--tw-translate-y))
+    rotate(var(--tw-rotate)) skewX(var(--tw-skew-x)) skewY(var(--tw-skew-y))
+    scaleX(var(--tw-scale-x)) scaleY(var(--tw-scale-y));
+}
+.SliderContainer [class="h-3 w-3 rounded-full bg-white dark:bg-gray-800"] {
+  background-color: #756aee !important;
+}
