+++ src/index.css
@@ -2,19 +2,4 @@
 @tailwind components;
 @tailwind utilities;
 
-::-webkit-scrollbar {
-  width: 5px;
-}
 
-::-webkit-scrollbar-thumb {
-  background: #756aee;
-  border-radius: 5px;
-}
-
-[data-testid="carousel-indicator"] {
-  background-color: #e2e8f0;
-  border-radius: 0px;
-}
-[class="h-3 w-3 rounded-full bg-white dark:bg-gray-800"] {
-  background-color: #756aee;
-}
